Mapped the command for intaking a note to the a button. moved the command files into the commands file.
