+++ src/main/java/frc/robot/commands/intake/Intake.java
@@ -2,15 +2,16 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.subsystems.Affector;
+package frc.robot.commands.intake;
 
 import edu.wpi.first.wpilibj2.command.Command;
+import frc.robot.subsystems.Intake.IntakeNote;
 
-public class Release extends Command {
-  /** Creates a new Release. */
+public class Intake extends Command {
+  /** Creates a new StartFlywheel. */
   private final IntakeNote intakeNote;
 
-  public Release(IntakeNote intakeNote) {
+  public Intake(IntakeNote intakeNote) {
     // Use addRequirements() here to declare subsystem dependencies.
     this.intakeNote = intakeNote;
     addRequirements(intakeNote);
@@ -23,7 +24,7 @@ public class Release extends Command {
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    intakeNote.release();
+    intakeNote.startIntake();
   }
 
   // Called once the command ends or is interrupted.
