+++ src/main/java/frc/robot/commands/intake/StopIntake.java
@@ -2,9 +2,10 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.subsystems.Affector;
+package frc.robot.commands.intake;
 
 import edu.wpi.first.wpilibj2.command.Command;
+import frc.robot.subsystems.Intake.IntakeNote;
 
 public class StopIntake extends Command {
   /** Creates a new Stop. */
