+++ src/main/java/frc/robot/subsystems/Intake/IntakeNoteIOMotorBrand.java
@@ -1,10 +1,7 @@
-package frc.robot.subsystems.Affector;
-
+package frc.robot.subsystems.Intake;
 
 // TODO: Implement setPercent() method with actual motor stuff when we know what motor we will use.
 
-/* Implements: To access the interface methods, the interface must be "implemented" 
+/* Implements: To access the interface methods, the interface must be "implemented"
 (kinda like inherited) by another class with the implements keyword (instead of extends). */
-public class IntakeNoteIOMotorBrand implements IntakeNoteIO {
-  
-}
+public class IntakeNoteIOMotorBrand implements IntakeNoteIO {}
